A-MoreOOP: Use more OOP
