+++ src/main/java/TaskList.java
@@ -1,0 +1,9 @@
+import java.util.ArrayList;
+import Task.*;
+
+
+public class TaskList extends ArrayList<Task> {
+    public TaskList() {
+        super();
+    }
+}
